requirements.txt
django==2.2
djangorestframework==3.10
psycopg2==2.8.3
requests==2.21.0